requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile pyproject.toml
#
aiofiles==23.2.1
    # via gradio
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   gradio
    #   httpx
    #   starlette
authlib==1.4.0
    # via gradiologin (pyproject.toml)
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   typer
    #   uvicorn
cryptography==44.0.0
    # via authlib
fastapi==0.115.6
    # via gradio
ffmpy==0.5.0
    # via gradio
filelock==3.16.1
    # via huggingface-hub
fsspec==2024.12.0
    # via
    #   gradio-client
    #   huggingface-hub
gradio==5.12.0
    # via gradiologin (pyproject.toml)
gradio-client==1.5.4
    # via gradio
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   gradio
    #   gradio-client
    #   safehttpx
huggingface-hub==0.27.1
    # via
    #   gradio
    #   gradio-client
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
itsdangerous==2.2.0
    # via gradiologin (pyproject.toml)
jinja2==3.1.5
    # via gradio
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   gradio
    #   jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==2.2.2
    # via
    #   gradio
    #   pandas
orjson==3.10.15
    # via gradio
packaging==24.2
    # via
    #   gradio
    #   gradio-client
    #   huggingface-hub
pandas==2.2.3
    # via gradio
pillow==11.1.0
    # via gradio
pycparser==2.22
    # via cffi
pydantic==2.10.5
    # via
    #   fastapi
    #   gradio
pydantic-core==2.27.2
    # via pydantic
pydub==0.25.1
    # via gradio
pygments==2.19.1
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-multipart==0.0.20
    # via gradio
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   gradio
    #   huggingface-hub
requests==2.32.3
    # via huggingface-hub
rich==13.9.4
    # via typer
ruff==0.9.2
    # via gradio
safehttpx==0.1.6
    # via gradio
semantic-version==2.10.0
    # via gradio
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio    # via

    #   fastapi
    #   gradio
tomlkit==0.13.2
    # via gradio
tqdm==4.67.1
    # via huggingface-hub
typer==0.15.1
    # via gradio
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   typer
tzdata==2024.2
    # via pandas
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via gradio
websockets==14.2
    # via gradio-client

google-auth==2.40.3
google-cloud==0.34.0
google-cloud-core==2.4.3
google-cloud-secret-manager==2.24.0
google-cloud-storage==3.3.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
cloud-sql-python-connector==1.18.4
google-api-core==2.25.1
SQLAlchemy==2.0.43
stripe==12.5.1
starlette==0.41.3
python-dotenv==1.0.0
pg8000>=1.30.0
